fix: address review feedback — import source, endpoint guard, test discoverability
- telemetry.test.ts: import InMemorySpanExporter/SimpleSpanProcessor from
  @opentelemetry/sdk-trace-node (listed dep) instead of sdk-trace-base
  (transitive only, would fail at runtime)
- telemetry.ts: add early-return guard when no OTLP endpoint is configured
  and no _testProvider is given, so initTelemetry truly is a no-op
- Move test files from src/__tests__/ to test/ so npm test glob finds them
- Switch test runner to tsx (handles .js→.ts resolution that
  --experimental-strip-types cannot, required by memory.ts → shared.ts)
- Revert tsconfig.json exclude change (tests no longer under src/)
