+++ package.json
@@ -43,7 +43,7 @@
     "build": "tsc",
     "dev": "tsc --watch",
     "start": "node dist/index.js",
-    "test": "node --test test/*.test.ts --experimental-strip-types"
+    "test": "npx tsx --test test/*.test.ts"
   },
   "engines": {
     "node": ">=20"
